# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: StreamFlix API - Netflix-style streaming & movie discovery platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/tmdb/import:
    post:
      operationId: importTmdbMovies
      tags: [admin]
      summary: Batch import movie metadata & trailers from TMDB API
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TmdbImportInput"
      responses:
        "200":
          description: Import results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TmdbImportResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/tmdb/refresh/{id}:
    post:
      operationId: refreshTmdbMovie
      tags: [admin]
      summary: Refresh metadata and trailer for a single movie from TMDB
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Movie refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Item:
      type: object
      required: [id, title, description, genres, tags, rating, releaseYear, createdAt]
      properties:
        id:
          type: integer
        tmdbId:
          type: ["integer", "null"]
        title:
          type: string
        tagline:
          type: ["string", "null"]
        description:
          type: string
        genres:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        rating:
          type: number
        voteCount:
          type: ["integer", "null"]
        popularity:
          type: ["number", "null"]
        releaseYear:
          type: integer
        releaseDate:
          type: ["string", "null"]
        posterUrl:
          type: ["string", "null"]
        backdropUrl:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        trailerUrl:
          type: ["string", "null"]
        trailerKey:
          type: ["string", "null"]
        trailerSite:
          type: ["string", "null"]
        trailerName:
          type: ["string", "null"]
        duration:
          type: ["integer", "null"]
        director:
          type: ["string", "null"]
        cast:
          type: array
          items:
            type: string
        originalLanguage:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        type:
          type: string
          enum: [movie, series, documentary]
        interactionCount:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
...
    TmdbImportInput:
      type: object
      required: [type, pages]
      properties:
        type:
          type: string
          enum: [popular, trending, top_rated, now_playing, genre, year]
        pages:
          type: integer
          default: 3
          minimum: 1
          maximum: 20
        genreId:
          type: integer
        genreName:
          type: string
        year:
          type: integer
    TmdbImportResponse:
      type: object
      required: [message, imported, updated, skipped, failed]
      properties:
        message:
          type: string
        imported:
          type: integer
        updated:
          type: integer
        skipped:
          type: integer
        failed:
          type: integer
        details:
          type: array
          items:
            type: string
...